実践力構築終わり

+++ WebContent/js/department/index.js
@@ -1,24 +1,60 @@
 function initialize() {
 	'use strict'
+
 	$.ajax({
 		type : 'GET',
-		url : '/EmployeesInfoManageTool/api/department/index',
+		url : '/EmployeesInfoManageTool/api/auth/self',
 		dataType: "json",
-		success: function(departmentList) {
-			const departmentTableBody = $("#department_table tbody")
-			for (let department of departmentList) {
-				const editLink = 'location.href="./edit.html?id=' + department.id + '&method=update"'
-				const deleteLink = 'location.href="./delete.html?id=' + department.id + '"'
-				const editButton = "<input type='button' onclick='" + editLink + "' value='編集' />"
-				const deleteButton = "<input type='button' department_id='" + department.id + "' onclick='departmentDeleteRequest(this)' value='削除' />"
-				const html = "<tr><td>" + department.id + "</td><td>" + department.name + "</td><td>" + editButton + "</td><td>" + deleteButton + "</td></tr>"
-				departmentTableBody.append(html)
+		statusCode: {
+			200: function(self) {
+				if (self.roleName !== "manager") disableManagerOperation()
+
+				$.ajax({
+					type : 'GET',
+					url : '/EmployeesInfoManageTool/api/department/index',
+					dataType: "json",
+					statusCode: {
+						200: function(departmentList) {
+							drawDepartmentTable(departmentList, self)
+						},
+						401: function() {
+							redirectLoginPage()
+						}
+					}
+				})
+			},
+			401: function() {
+				redirectLoginPage()
 			}
 		}
 	})
 }
 
-function departmentDeleteRequest(button) {
+function drawDepartmentTable(departmentList, self) {
+	const departmentTableBody = $("#department_table tbody")
+	for (let department of departmentList) {
+		let html = "<tr><td>" + department.id + "</td><td>" + department.name + "</td>"
+
+		if (self.roleName === "manager") {
+			const editLink = 'location.href="./edit.html?id=' + department.id + '"'
+			const deleteLink = 'location.href="./delete.html?id=' + department.id + '"'
+			const editButton = "<input type='button' onclick='" + editLink + "' value='編集' />"
+			const deleteButton = "<input type='button' department_id='" + department.id + "' onclick='deleteDepartmentRequest(this)' value='削除' />"
+			html += "<td>" + editButton + "</td><td>" + deleteButton + "</td>"
+		}
+
+		html += "</tr>"
+
+		departmentTableBody.append(html)
+	}
+}
+
+function redirectLoginPage() {
+	alert("ログインしてください")
+	location.href = "/EmployeesInfoManageTool/html/authentication/login.html"
+}
+
+function deleteDepartmentRequest(button) {
 	'use strict'
 
 	const departmentId = $(button).attr("department_id")
@@ -28,17 +64,26 @@ function departmentDeleteRequest(button) {
 		url : '/EmployeesInfoManageTool/api/department/delete',
 		dataType: "json",
 		data: {id: departmentId},
-		success: function(result) {
-			if (result === true) {
-				location.reload()
-			} else {
-				$("#message").html("削除に失敗しました")
+		statusCode: {
+			200: function(result) {
+				if (result === true) {
+					location.reload()
+				} else {
+					$("#message").html("削除に失敗しました")
+				}
+			},
+			401: function() {
+				redirectLoginPage()
+			},
+			403: function() {
+				alert("権限がありません")
 			}
-		},
-		error:function(XMLHttpRequest, textStatus, errorThrown) {
-			$("#message").html("削除に失敗しました")
 		}
 	})
+}
+
+function disableManagerOperation() {
+	$(".manager").attr("hidden", true)
 }
 
 $(document).ready(function () {
