+++ WebContent/js/employee/index.js
@@ -1,24 +1,68 @@
 function initialize() {
 	'use strict'
+
 	$.ajax({
 		type : 'GET',
-		url : '/EmployeesInfoManageTool/api/employee/index',
+		url : '/EmployeesInfoManageTool/api/auth/self',
 		dataType: "json",
-		success: function(employeeList) {
-			const tBody = $("#employee_list")
-			for (let employee of employeeList) {
-				const editLink = 'location.href="./edit.html?id=' + employee.id + '&method=update"'
-				const deleteLink = 'location.href="./delete.html?id=' + employee.id + '"'
-				const editButton = "<input type='button' onclick='" + editLink + "' value='編集' />"
-				const deleteButton = "<input type='button' employee_id='" +employee.id + "' onclick='employeeDeleteRequest(this)' value='削除' />"
-				const html = "<tr><td>" + employee.id + "</td><td>" + employee.employeeDetail.name + "</td><td>" + editButton + "</td><td>" + deleteButton + "</td></tr>"
-				tBody.append(html)
+		statusCode: {
+			200: function(self) {
+				if (self.roleName !== "manager") disableManagerOperation()
+
+				$.ajax({
+					type : 'GET',
+					url : '/EmployeesInfoManageTool/api/employee/index',
+					dataType: "json",
+					statusCode: {
+						200: function(employeeList) {
+							drawEmployeeTable(employeeList, self)
+						},
+						401: function() {
+							redirectLoginPage()
+						}
+					}
+				})
+			},
+			401: function() {
+				redirectLoginPage()
 			}
 		}
 	})
 }
 
-function employeeDeleteRequest(button) {
+function redirectLoginPage() {
+	alert("ログインしてください")
+	location.href = "/EmployeesInfoManageTool/html/auth/login.html"
+}
+
+function disableManagerOperation() {
+	$(".manager").attr("hidden", true)
+}
+
+function drawEmployeeTable(employeeList, self) {
+	const employeeTableBody = $("#employee_table tbody")
+	for (let employee of employeeList) {
+		let html = "<tr><td>" + employee.id + "</td><td>" + employee.employeeDetail.name + "</td>"
+
+		if (self.roleName === "manager") {
+			const editLink = 'location.href="./edit.html?id=' + employee.id + '&method=update"'
+			const editButton = "<input type='button' onclick='" + editLink + "' value='編集' />"
+			const deleteLink = 'location.href="./delete.html?id=' + employee.id + '"'
+			const deleteButton = "<input type='button' employee_id='" +employee.id + "' onclick='deleteEmployeeRequest(this)' value='削除' />"
+			html += "<td>" + editButton + "</td><td>" + deleteButton + "</td>"
+		} else if (self.employeeId === employee.id) {
+			const editLink = 'location.href="./edit.html?id=' + employee.id + '&method=update"'
+			const editButton = "<input type='button' onclick='" + editLink + "' value='編集' />"
+			html += "<td>" + editButton + "</td>"
+		}
+
+		html += "</tr>"
+
+		employeeTableBody.append(html)
+	}
+}
+
+function deleteEmployeeRequest(button) {
 	'use strict'
 
 	const employeeId = $(button).attr("employee_id")
@@ -28,18 +72,21 @@ function employeeDeleteRequest(button) {
 		url : '/EmployeesInfoManageTool/api/employee/delete',
 		dataType: "json",
 		data: {id: employeeId},
-		success: function(result) {
-			console.log("test")
-			if (result === true) {
-				location.reload()
-			} else {
-				$("#message").html("削除に失敗しました")
+		statusCode: {
+			200: function(result) {
+				if (result === true) {
+					location.reload()
+				} else {
+					$("#message").html("削除に失敗しました")
+				}
+			},
+			401: function() {
+				redirectLoginPage()
+			},
+			403: function() {
+				alert("権限がありません")
 			}
-		},
-		error:function(XMLHttpRequest, textStatus, errorThrown) {
-			$("#message").html("削除に失敗しました")
 		}
-
 	})
 }
